Vote_Prediction_ML_Project
📊 Project: Vote Prediction

🔍 Description:
As part of a collaboration with CNBE, a prominent news channel, I undertook a comprehensive analysis of recent elections. The survey data, comprising responses from 1525 voters across 9 variables, served as the foundation for this predictive modeling project.

🗳️ Objective:
The primary objective was to develop a robust predictive model capable of forecasting the political party preference of individual voters based on the provided information. By leveraging advanced machine learning techniques, the aim was to create an exit poll tool to aid in predicting the overall election outcome and the distribution of seats among different political parties.

💡 Approach:
Utilizing the survey data, I applied various machine learning algorithms to train and fine-tune the predictive model. Feature engineering, data preprocessing, and model selection were integral steps in the process, ensuring the accuracy and reliability of the predictions. The ultimate goal was to provide CNBE with actionable insights to facilitate informed analysis and reporting of election results.

📈 Impact:
The developed predictive model not only enabled accurate individual-level predictions but also contributed to the broader understanding of voter behavior and electoral dynamics. By facilitating the creation of an effective exit poll, the project played a crucial role in shaping public discourse and enhancing the transparency of the democratic process.